+++ settings.gradle
@@ -25,7 +25,6 @@ include ':simplified-books-database'
 include ':simplified-books-database-api'
 include ':simplified-books-registry-api'
 include ':simplified-boot-api'
-include ':simplified-branding'
 include ':simplified-bugsnag'
 include ':simplified-cardcreator'
 include ':simplified-datepicker'
@@ -71,4 +70,5 @@ include ':simplified-tests-sandbox'
 include ':simplified-tests-strings'
 include ':simplified-theme'
 include ':simplified-threads'
+include ':simplified-ui-branding'
