+++ simplified-app-vanilla-with-profiles/src/main/resources/META-INF/services/org.nypl.simplified.ui.branding.BrandingSplashServiceType
