+++ simplified-app-vanilla/src/main/java/org/nypl/simplified/vanilla/BrandingThemeService.kt
@@ -1,6 +1,6 @@
 package org.nypl.simplified.vanilla
 
-import org.nypl.simplified.branding.BrandingThemeOverrideServiceType
+import org.nypl.simplified.ui.branding.BrandingThemeOverrideServiceType
 import org.nypl.simplified.theme.ThemeValue
 
 class BrandingThemeService : BrandingThemeOverrideServiceType {
