+++ simplified-app-vanilla/src/main/resources/META-INF/services/org.nypl.simplified.ui.branding.BrandingThemeOverrideServiceType
