+++ simplified-app-vanilla/version.properties
@@ -1,4 +1,4 @@
 #
-#Mon Oct 07 18:04:51 UTC 2019
+#Thu Oct 10 09:36:54 UTC 2019
 versionName=3.0.14
-versionCode=4156
+versionCode=4160
