+++ simplified-tests/build.gradle
@@ -21,7 +21,6 @@ dependencies {
   api project(":simplified-books-database-api")
   api project(":simplified-books-registry-api")
   api project(":simplified-boot-api")
-  api project(":simplified-branding")
   api project(":simplified-accounts-source-spi")
 
   api(project(":simplified-bugsnag")) {
@@ -58,6 +57,7 @@ dependencies {
   api project(":simplified-tests-strings")
   api project(":simplified-theme")
   api project(":simplified-threads")
+  api project(":simplified-ui-branding")
 
   api libraries.nyplAudioBookAPI
   api libraries.nyplAudioBookManifestNYPL
