+++ simplified-ui-branding/gradle.properties
